feat: add solutions to modules
